+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Version schema is `year.month.num_release`
 ### Added
 
 - ZTF DR13 is added. Note that DR8 is still the default, since we do not have DR13 UCI installation yet
+- SPICY catalog cross-matching ([Vizier](https://vizier.cds.unistra.fr/viz-bin/VizieR?-source=J/ApJS/254/33)) https://github.com/snad-space/ztf-viewer/pull/233 https://github.com/snad-space/ztf-viewer/pull/234
 
 ### Fixed